Testing VOPR swarm testing with better fault selection strategy.
Now, we have a top-level switch that decides whether swarm testing should
be enabled or not, a switch at the level of
each class of faults (network, storage, and process) that decides whether
that entire class of faults should be disabled, and a switch at
the level of each sub-fault (for example, partitions under network),
that decies whether that sub-fault should be disabled. All of these are controlled
using random.boolean().